pyproject.toml
[tool.poetry]
name = "devncode"
version = "0.1.0"
description = ""
authors = ["Marco Cotrufo <marco.cotrufo@devncode.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
python-slugify = "^8.0.4"
python-frontmatter = "^1.1.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
black = "^24.10.0"
mypy = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
